满意度去除小数点，增加百分号

+++ app/src/main/java/com/floyd/diamond/bean/MoteDetail1.java
@@ -32,7 +32,7 @@ public class MoteDetail1 {
     public int gender;
     public int orderNum;
     private int shape;
-    public float goodeEvalRate;
+    public int goodeEvalRate;
 
     public void setIsFollow(boolean isFollow) {
         this.isFollow = isFollow;
