+++ app/src/main/java/com/floyd/diamond/biz/vo/mote/MoteInfoVO.java
@@ -16,7 +16,7 @@ public class MoteInfoVO implements IKeepClassForProguard {
     public int orderNum;
     public int fenNum;
     public int followNum;
-    public float goodeEvalRate;
+    public int goodeEvalRate;
     public int fee;
     public String authenPic1;//验证图片1
     public String authenPic2;//验证图片2
